ROyFit

ROyFit

NeRCA.svdfit — Method.

function svdfit(hits::Vector{CalibratedHit})

Uses SVD to do a fast and dirty track prefit. Provide hits with a multiplicity of at least 2.

source

NeRCA.make_cherenkov_calculator — Method.

function make_cherenkov_calculator(d_closest, t_closest, z_closest, dir_z, t₀)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

source

NeRCA.make_cherenkov_calculator — Method.

function make_cherenkov_calculator(sdp::SingleDUParams)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

source

NeRCA.make_cherenkov_calculator — Method.

function make_cherenkov_calculator(track::Track, event_info::Union{MCEventInfo,DAQEventInfo}; v=2.99792458e8)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

source

NeRCA.make_cherenkov_calculator — Method.

function make_cherenkov_calculator(track::Track; v=2.99792458e8)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

source

NeRCA.prefit — Method.

function prefit(hits::Vector{CalibratedHit})

Performs the prefit algorithm which was used in DUMAND II.

source

NeRCA.single_du_params — Method.

function single_du_params(t::NeRCA.MCTrack)

Calculates five parameters to describe a MC track for a single DU case.

source

NeRCA.single_du_params — Method.

function single_du_params(track::NeRCA.Track)

Calculates five parameters to describe a track for a single DU case.

source

NeRCA.slew — Method.

function slew(tot)

Return the time slewing for a ToT.

source