Working with Hits

The hit type hierarchy is currently redesigned!

NeRCA.jl has two main types of hits: AbstractDAQHits and AbstractMCHits.

NeRCA.domhits — Method

function domhits(hits::Vector{T}) where {T<:Hit}

Sort hits by DOM ID and put them into a dictionary.

NeRCA.duhits — Method

function duhits(hits::Vector{T}) where {T<:CalibratedHit}

Sort hits by DU and put them into a dictionary.

NeRCA.nfoldhits — Method

function nfoldhits(hits::Vector{T}, Δt, n) where {T<:Hit}

Create a Vector with hits contributing to n-fold coincidences within a time window of Δt.

NeRCA.triggered — Method

function triggered(hits::Vector{T}) where {T<:Hit}

Return a Vector of triggered hits.

NeRCA.count_multiplicities! — Function

function count_multiplicities!(hits::Vector{CalibratedHit}, tmax=20)

Counts the multiplicities and modifies the .multiplicity field of the hits. Important: the hits have to be sorted by time and then by DOM ID first.

NeRCA.count_multiplicities — Method

function count_multiplicities(hits::Vector{T}, tmax=20) where {T<:AbstractHit}

Calculate the multiplicities for a given time window. Two arrays are are returned, one contains the multiplicities, the second one the IDs of the coincidence groups. The hits should be sorted by time and then by dom_id.

NeRCA.nphes — Method

function nphes(tot)

Returns the estimated number of photoelectrons for a given ToT.

NeRCA.slew — Method

function slew(tot)

Return the time slewing for a ToT.

NeRCA.totcut — Method

function totcut(hits::Vector{T}, tot) where {T<:DAQHit}

Return a vector of hits with ToT >= tot.