ROyFit

NeRCA.make_cherenkov_calculator — Method

make_cherenkov_calculator(d_closest, t_closest, z_closest, dir_z, t₀; n)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

NeRCA.make_cherenkov_calculator — Method

make_cherenkov_calculator(track, event_info; v)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

NeRCA.make_cherenkov_calculator — Method

make_cherenkov_calculator(track; v, n)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

NeRCA.make_cherenkov_calculator — Method

make_cherenkov_calculator(sdp; n)

Returns a function which calculates the arrival time of a Cherenkov photon at a given position.

NeRCA.prefit — Method

prefit(hits)

Performs the prefit algorithm which was used in DUMAND II.

NeRCA.select_hits — Method

select_hits(du_hits, hit_pool; Δt₋, Δz, new_hits)

Returns the seed hits suited for a Cherenkov hit time residual based reconstruction algorithm.

The du_hits should only contain hits for a single DU. The hitpool holds all other hit candidates (e.g. created by `createhit_pool()).Δt₋is the allowed negative time error for the arrival time,Δz` distance between two floors.

NeRCA.single_du_params — Method

single_du_params(t, time)

Calculates five parameters to describe a MC track for a single DU case.

NeRCA.single_du_params — Method

single_du_params(track)

Calculates five parameters to describe a track for a single DU case.