Hits

The hit type hierarchy is currently being redesigned!

NeRCA.jl has three main types of hits: AbstractDAQHit, AbstractMCHit and CalibratedHit.

hit type	supertype	guaranteed attributes

Hit (in future SnapshotHit)	AbstractDAQHit	channel_id, dom_id, t, tot

TriggeredHit	AbstractDAQHit	channel_id, dom_id, t, tot, trigger_mask

MCHit	AbstractMCHit	a, origin, pmt_id, t

NeRCA.combine — Method

combine(snapshot_hits, triggered_hits)

Combine snapshot and triggered hits to a single hits-vector.

This should be used to transfer the trigger information to the snapshot hits from a DAQEvent. The triggered hits are a subset of the snapshot hits.

NeRCA.count_multiplicities! — Function

count_multiplicities!(hits)
count_multiplicities!(hits, tmax)

Counts the multiplicities and modifies the .multiplicity field of the hits. Important: the hits have to be sorted by time and then by DOM ID first.

NeRCA.count_multiplicities — Method

count_multiplicities(hits)
count_multiplicities(hits, tmax)

Calculate the multiplicities for a given time window. Two arrays are are returned, one contains the multiplicities, the second one the IDs of the coincidence groups. The hits should be sorted by time and then by dom_id.

NeRCA.domhits — Method

domhits(hits)

Sort hits by DOM ID and put them into a dictionary.

NeRCA.duhits — Method

duhits(hits)

Sort hits by DU and put them into a dictionary.

NeRCA.nfoldhits — Method

nfoldhits(hits, Δt, n)

Create a Vector with hits contributing to n-fold coincidences within a time window of Δt.

NeRCA.nphes — Method

nphes(tot)

Returns the estimated number of photoelectrons for a given ToT.

NeRCA.totcut — Method

totcut(hits, tot)

Return a vector of hits with ToT >= tot.

NeRCA.triggered — Method

triggered(hits)

Return only triggered hits.